+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 
 * Moved `*Task` classes to `lib/thermite/rails/tasks/`.
+* Added rake task for running `cargo update` for all crates.
 
 ## [0.1.0] - 2018-03-27